feat(utils): add new package and consolidate test utils
