+++ packages/eslint-plugin/tests/rules/no-input-prefix.test.ts
@@ -1,8 +1,8 @@
-import rule, { MessageIds, RULE_NAME } from '../../src/rules/no-input-prefix';
 import {
   convertAnnotatedSourceToFailureCase,
   RuleTester,
-} from '../test-helper';
+} from '@angular-eslint/utils';
+import rule, { MessageIds, RULE_NAME } from '../../src/rules/no-input-prefix';
 
 //------------------------------------------------------------------------------
 // Tests
