+++ packages/eslint-plugin/tests/rules/no-output-native.test.ts
@@ -1,8 +1,8 @@
-import rule, { MessageIds, RULE_NAME } from '../../src/rules/no-output-native';
 import {
   convertAnnotatedSourceToFailureCase,
   RuleTester,
-} from '../test-helper';
+} from '@angular-eslint/utils';
+import rule, { MessageIds, RULE_NAME } from '../../src/rules/no-output-native';
 
 //------------------------------------------------------------------------------
 // Tests
